third_party/blink/web_tests/wpt_internal/webmidi/requestmidiaccess-basic.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
assert_implements(navigator.requestMIDIAccess);

test(() => {
  const result = navigator.requestMIDIAccess();
  assert_true(result instanceof Promise);
}, 'requestMIDIAccess is defined and callable');
</script>
</body>
</html>
